+++ src/domains/PostDetail/components/PostDetailTitle/PostDetailTitle.style.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const PostDetailTitleContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  padding: 10px 20px;
+
+  h2 {
+    font-size: 18px;
+    font-weight: var(--weight-semi-bold);
+
+    margin-bottom: 5px;
+  }
+`;
